Implement comparing more than two histograms (averaging-comparison). Create start script, and add usage info to the README.

+++ src/main/java/com/trivago/jcha/ClassHistogramStats.java
@@ -52,7 +52,7 @@ public class ClassHistogramStats
 				continue;
 			}
 
-			int byteDiffAbsolute = Math.abs(chse.byteDiff);
+			long byteDiffAbsolute = Math.abs(chse.byteDiff);
 			if (byteDiffAbsolute < minByteDiff)
 			{
 				continue;
