+++ src/main/java/com/trivago/jcha/comparator/AbsoluteSizeComparator.java
@@ -1,7 +1,5 @@
 package com.trivago.jcha.comparator;
 
-import java.util.Comparator;
-
 import com.trivago.jcha.ClassHistogramStatsEntry;
 
 
@@ -17,7 +15,7 @@ public class AbsoluteSizeComparator extends BaseComparator
 	@Override
 	public int compare(ClassHistogramStatsEntry o1, ClassHistogramStatsEntry o2)
 	{
-		int diff = o2.getByteDiff() - o1.getByteDiff(); 
+		long diff = o2.getByteDiff() - o1.getByteDiff(); 
 		return compareBase(o1, o2, diff);
 	}
